fix(contexts): rename MenuModeContext.tsx to MenuModeProvider.tsx
The provider file and the context-object file (MenuModeContext.tsx and
menuModeContext.ts) only differed by leading-letter case. On macOS's
case-insensitive filesystem TypeScript treated them as the same module
and resolved the App.tsx import to the wrong one, surfacing TS2305 and
TS1149/TS1261 errors. Renaming the JSX provider to a truly distinct
name (MenuModeProvider.tsx) keeps the file-per-concern split that
React Fast Refresh wants without the casing collision.
